+++ include/private/formatter.h
@@ -23,21 +23,17 @@
 #  include <stumpless/target.h>
 #  include "private/strbuilder.h"
 
+#  define RFC_5424_FULL_DATE_BUFFER_SIZE 11
+#  define RFC_5424_FULL_TIME_BUFFER_SIZE 10
 #  define RFC_5424_MAX_PRI_LENGTH 5
 #  define RFC_5424_MAX_TIMESTAMP_LENGTH 32
 #  define RFC_5424_MAX_HOSTNAME_LENGTH 255
 #  define RFC_5424_MAX_PROCID_LENGTH 128
+#  define RFC_5424_TIME_SECFRAC_BUFFER_SIZE 8
+#  define RFC_5424_TIMESTAMP_BUFFER_SIZE 33
+#  define RFC_5424_WHOLE_TIME_BUFFER_SIZE 20
 
 struct strbuilder *
-format_entry( const struct stumpless_target *target,
-              struct stumpless_entry *entry );
-
-/*
- * Gets the current timestamp and writes it to the string builder. The time
- * reflected by the timestamp will be as close to the time of the function
- * invocation as feasible.
- */
-struct strbuilder *
-strbuilder_append_rfc5424_timestamp( struct strbuilder *builder );
+format_entry( struct stumpless_entry *entry );
 
 #endif /* __STUMPLESS_PRIVATE_FORMATTER_H */